pyproject.toml
[project]
name = "fastapi__allnotes"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "saxetr", email = "86871918+saxetr@users.noreply.github.com"},
]
dependencies = [
    "annotated-types==0.6.0",
    "anyio==4.3.0",
    "asttokens==2.4.1",
    "certifi==2024.2.2",
    "click==8.1.7",
    "decorator==5.1.1",
    "dnspython==2.6.1",
    "email-validator==2.1.1",
    "executing==2.0.1",
    "fastapi==0.110.1",
    "greenlet==3.0.3",
    "h11==0.14.0",
    "httpcore==1.0.5",
    "httptools==0.6.1",
    "httpx==0.27.0",
    "idna==3.7",
    "iniconfig==2.0.0",
    "ipython==8.23.0",
    "itsdangerous==2.1.2",
    "jedi==0.19.1",
    "jinja2>=3.1.3",
    "markdown-it-py==3.0.0",
    "MarkupSafe==2.1.5",
    "matplotlib-inline==0.1.6",
    "mdurl==0.1.2",
    "orjson==3.10.0",
    "packaging==24.0",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pluggy==1.4.0",
    "pprintpp==0.4.0",
    "prompt-toolkit==3.0.43",
    "psycopg2-binary==2.9.9",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.2",
    "pydantic==2.7.0",
    "pydantic-extra-types==2.6.0",
    "pydantic-settings==2.2.1",
    "pydantic-core==2.18.1",
    "Pygments==2.17.2",
    "pytest==8.1.1",
    "pytest-clarity==1.0.1",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "PyYAML==6.0.1",
    "rich==13.7.1",
    "shellingham==1.5.4",
    "six==1.16.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.29",
    "stack-data==0.6.3",
    "starlette==0.37.2",
    "traitlets==5.14.2",
    "typer==0.12.3",
    "typing-extensions==4.11.0",
    "ujson==5.9.0",
    "uvicorn==0.29.0",
    "uvloop==0.19.0",
    "watchfiles==0.21.0",
    "wcwidth==0.2.13",
    "websockets==12.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
test = [
    "pytest>=8.1.1",
    "pytest-clarity>=1.0.1",
    "ipython>=8.23.0",
]
